Rename resources -> resource
