+++ src/main/java/net/neoforged/neoforge/transfer/DelegatingResourceHandler.java
@@ -7,7 +7,7 @@ package net.neoforged.neoforge.transfer;
 
 import java.util.Objects;
 import java.util.function.Supplier;
-import net.neoforged.neoforge.transfer.resources.IResource;
+import net.neoforged.neoforge.transfer.resource.IResource;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 
 /**
