+++ src/main/java/net/neoforged/neoforge/transfer/EmptyResourceHandler.java
@@ -5,7 +5,7 @@
 
 package net.neoforged.neoforge.transfer;
 
-import net.neoforged.neoforge.transfer.resources.IResource;
+import net.neoforged.neoforge.transfer.resource.IResource;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 
 /**
