+++ src/main/java/net/neoforged/neoforge/transfer/fluid/FluidResource.java
@@ -25,7 +25,7 @@ import net.minecraft.world.level.material.Fluids;
 import net.neoforged.neoforge.fluids.FluidStack;
 import net.neoforged.neoforge.fluids.FluidType;
 import net.neoforged.neoforge.transfer.TransferPreconditions;
-import net.neoforged.neoforge.transfer.resources.IDataComponentHolderResource;
+import net.neoforged.neoforge.transfer.resource.IDataComponentHolderResource;
 
 /**
  * Immutable combination of a {@link Fluid} and data components.
