+++ src/main/java/net/neoforged/neoforge/transfer/item/ItemResource.java
@@ -25,7 +25,7 @@ import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.Items;
 import net.minecraft.world.level.ItemLike;
 import net.neoforged.neoforge.transfer.TransferPreconditions;
-import net.neoforged.neoforge.transfer.resources.IDataComponentHolderResource;
+import net.neoforged.neoforge.transfer.resource.IDataComponentHolderResource;
 
 /**
  * Immutable combination of an {@link Item} and data components.
