+++ src/main/java/net/neoforged/neoforge/transfer/resource/IDataComponentHolderResource.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.transfer.resources;
+package net.neoforged.neoforge.transfer.resource;
 
 import java.util.function.Supplier;
 import net.minecraft.core.component.DataComponentHolder;
