+++ src/main/java/net/neoforged/neoforge/transfer/resource/ResourceStack.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: LGPL-2.1-only
  */
 
-package net.neoforged.neoforge.transfer.resources;
+package net.neoforged.neoforge.transfer.resource;
 
 import net.neoforged.neoforge.transfer.ResourceHandlerUtil;
