+++ tests/src/junit/java/net/neoforged/neoforge/unittest/transfer/ResourceHandlerUtilTest.java
@@ -23,7 +23,7 @@ import net.minecraft.util.Mth;
 import net.minecraft.world.level.redstone.Redstone;
 import net.neoforged.neoforge.transfer.EmptyResourceHandler;
 import net.neoforged.neoforge.transfer.ResourceHandlerUtil;
-import net.neoforged.neoforge.transfer.resources.ResourceStack;
+import net.neoforged.neoforge.transfer.resource.ResourceStack;
 import net.neoforged.neoforge.transfer.transaction.Transaction;
 import net.neoforged.neoforge.transfer.transaction.TransactionContext;
 import org.junit.jupiter.api.Nested;
